# human_agent/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Human Agent",
    "description": "Read-only world v0.12 analysis, never a world step API.",
    "version": "2.0"
  },
  "paths": {
...
    "/human-agent/analyze": {
      "post": {
        "summary": "Analyze",
        "operationId": "analyze_human_agent_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisOptions": {
        "properties": {
          "include_recommendations": {
            "type": "boolean",
            "title": "Include Recommendations",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AnalysisOptions"
      },
      "AnalyzeRequest": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "2.0",
            "title": "Schema Version"
          },
          "request_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Request Id"
          },
          "state_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "State Id"
          },
          "world_rules_version": {
            "type": "string",
            "const": "0.12",
            "title": "World Rules Version"
          },
          "tick": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Tick"
          },
          "snapshot_phase": {
            "type": "string",
            "const": "between_ticks",
            "title": "Snapshot Phase"
          },
          "world_status": {
            "type": "string",
            "enum": [
              "running",
              "paused",
              "planning",
              "error_paused",
              "failed"
            ],
            "title": "World Status"
          },
          "resources": {
            "$ref": "#/components/schemas/Resources"
          },
          "crew": {
            "items": {
              "$ref": "#/components/schemas/Crew"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Crew"
          },
          "plots": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Plot"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plots"
          },
          "water_production_available": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Water Production Available"
          },
          "next_tick_plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Plan"
              },
              {
                "type": "null"
              }
            ]
          },
          "analysis": {
            "$ref": "#/components/schemas/AnalysisOptions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "schema_version",
          "request_id",
          "state_id",
          "world_rules_version",
          "tick",
          "snapshot_phase",
          "world_status",
          "resources",
          "crew"
        ],
        "title": "AnalyzeRequest"
      },
      "AnalyzeResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "2.0",
            "title": "Schema Version"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "state_id": {
            "type": "string",
            "title": "State Id"
          },
          "tick": {
            "type": "integer",
            "title": "Tick"
          },
          "for_tick": {
            "type": "integer",
            "title": "For Tick"
          },
          "world_rules_version": {
            "type": "string",
            "const": "0.12",
            "title": "World Rules Version"
          },
          "rules_hash": {
            "type": "string",
            "title": "Rules Hash"
          },
          "analysis_status": {
            "type": "string",
            "enum": [
              "complete",
              "partial"
            ],
            "title": "Analysis Status"
          },
          "execution_mode": {
            "type": "string",
            "enum": [
              "live",
              "mock",
              "degraded",
              "deterministic"
            ],
            "title": "Execution Mode"
          },
          "current_world_condition": {
            "type": "string",
            "enum": [
              "active",
              "already_failed",
              "inconsistent",
              "unknown"
            ],
            "title": "Current World Condition"
          },
          "crew_assessments": {
            "items": {
              "$ref": "#/components/schemas/CrewAssessment"
            },
            "type": "array",
            "title": "Crew Assessments"
          },
          "public_resource_assessment": {
            "additionalProperties": true,
            "type": "object",
            "title": "Public Resource Assessment"
          },
          "workforce_summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Workforce Summary"
          },
          "next_tick_audit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Audit"
              },
              {
                "type": "null"
              }
            ]
          },
          "risks": {
            "items": {
              "$ref": "#/components/schemas/Risk"
            },
            "type": "array",
            "title": "Risks"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/Recommendation"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "human_requests_to_core": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Human Requests To Core"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence"
          },
          "retrieval": {
            "$ref": "#/components/schemas/RetrievalInfo"
          },
          "corpus_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corpus Version"
          },
          "index_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Index Version"
          },
          "research_status": {
            "type": "string",
            "enum": [
              "complete",
              "partial"
            ],
            "title": "Research Status"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assumptions"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "missing_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Fields"
          },
          "integration_gaps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Integration Gaps"
          },
          "diagnostics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Diagnostics"
          },
          "tool_trace": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Tool Trace"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "schema_version",
          "request_id",
          "state_id",
          "tick",
          "for_tick",
          "world_rules_version",
          "rules_hash",
          "analysis_status",
          "execution_mode",
          "current_world_condition",
          "crew_assessments",
          "public_resource_assessment",
          "workforce_summary",
          "next_tick_audit",
          "risks",
          "recommendations",
          "human_requests_to_core",
          "evidence",
          "retrieval",
          "corpus_version",
          "index_version",
          "research_status",
          "assumptions",
          "warnings",
          "missing_fields",
          "integration_gaps",
          "diagnostics",
          "tool_trace"
        ],
        "title": "AnalyzeResponse"
      },
      "Audit": {
        "properties": {
          "result_id": {
            "type": "string",
            "title": "Result Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "feasible_in_scope",
              "fatal_in_scope",
              "unsafe_in_scope",
              "incomplete",
              "not_applicable"
            ],
            "title": "Status"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "for_tick": {
            "type": "integer",
            "title": "For Tick"
          },
          "coverage_end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage End"
          },
          "fatal_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fatal Stage"
          },
          "stage_results": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Stage"
            },
            "type": "object",
            "title": "Stage Results"
          },
          "ledger": {
            "items": {
              "$ref": "#/components/schemas/LedgerEntry"
            },
            "type": "array",
            "title": "Ledger"
          },
          "known_after_values": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Known After Values"
          },
          "fatal_conditions": {
            "items": {
              "$ref": "#/components/schemas/FatalCondition"
            },
            "type": "array",
            "title": "Fatal Conditions"
          },
          "unknown_dependencies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unknown Dependencies"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "result_id",
          "status",
          "scope",
          "for_tick",
          "coverage_end",
          "fatal_stage",
          "stage_results",
          "ledger",
          "known_after_values",
          "fatal_conditions",
          "unknown_dependencies",
          "limitations"
        ],
        "title": "Audit"
      },
...
      "Crew": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Id"
          },
          "alive": {
            "type": "boolean",
            "title": "Alive"
          },
          "food_energy": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 3000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Food Energy"
          },
          "water": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 2.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Water"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "alive",
          "food_energy",
          "water"
        ],
        "title": "Crew"
      },
      "CrewAssessment": {
        "properties": {
          "crew_id": {
            "type": "string",
            "title": "Crew Id"
          },
          "alive": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alive"
          },
          "food_energy": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Food Energy"
          },
          "water": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Water"
          },
          "energy_ratio": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Energy Ratio"
          },
          "water_ratio": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Water Ratio"
          },
          "low_flags": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Low Flags"
          },
          "idle_no_refill": {
            "$ref": "#/components/schemas/Margins"
          },
          "requested_work_no_refill": {
            "$ref": "#/components/schemas/Margins"
          },
          "after_refill": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "number",
                      "minimum": 0.0
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "After Refill"
          },
          "after_base": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "number",
                      "minimum": 0.0
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "After Base"
          },
          "actual_work": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Work"
          },
          "fatal_conditions": {
            "items": {
              "$ref": "#/components/schemas/FatalCondition"
            },
            "type": "array",
            "title": "Fatal Conditions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "crew_id",
          "alive",
          "food_energy",
          "water",
          "energy_ratio",
          "water_ratio",
          "low_flags",
          "idle_no_refill",
          "requested_work_no_refill",
          "after_refill",
          "after_base",
          "actual_work",
          "fatal_conditions"
        ],
        "title": "CrewAssessment"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Plan": {
        "properties": {
          "for_tick": {
            "type": "integer",
            "minimum": 1.0,
            "title": "For Tick"
          },
          "crew_tasks": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Idle"
                },
                {
                  "$ref": "#/components/schemas/Eat"
                },
                {
                  "$ref": "#/components/schemas/Drink"
                },
                {
                  "$ref": "#/components/schemas/Generate"
                },
                {
                  "$ref": "#/components/schemas/CropTask"
                },
                {
                  "$ref": "#/components/schemas/Plant"
                }
              ],
              "discriminator": {
                "propertyName": "task",
                "mapping": {
                  "clear": "#/components/schemas/CropTask",
                  "drink": "#/components/schemas/Drink",
                  "eat": "#/components/schemas/Eat",
                  "generate": "#/components/schemas/Generate",
                  "harvest": "#/components/schemas/CropTask",
                  "idle": "#/components/schemas/Idle",
                  "plant": "#/components/schemas/Plant"
                }
              }
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Crew Tasks"
          },
          "refill_order": {
            "items": {
              "type": "string",
              "maxLength": 128,
              "minLength": 1
            },
            "type": "array",
            "title": "Refill Order"
          },
          "water_production_request_liters": {
            "type": "number",
            "minimum": 0.0,
            "title": "Water Production Request Liters"
          },
          "irrigation_allocations": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Allocation"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Irrigation Allocations"
          },
          "crop_operation_order": {
            "items": {
              "type": "string",
              "maxLength": 128,
              "minLength": 1
            },
            "type": "array",
            "title": "Crop Operation Order"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "for_tick",
          "crew_tasks",
          "refill_order",
          "water_production_request_liters",
          "irrigation_allocations",
          "crop_operation_order"
        ],
        "title": "Plan"
      },
...
      "Plot": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "empty",
              "growing",
              "mature",
              "dead"
            ],
            "title": "Status"
          },
          "crop_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "lettuce",
                  "potato",
                  "tomato",
                  "wheat",
                  "soybean"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Crop Type"
          },
          "growth_ticks": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Growth Ticks"
          },
          "consecutive_unirrigated_ticks": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Consecutive Unirrigated Ticks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "status",
          "crop_type",
          "growth_ticks",
          "consecutive_unirrigated_ticks"
        ],
        "title": "Plot"
      },
...
      "Recommendation": {
        "properties": {
          "proposal_id": {
            "type": "string",
            "title": "Proposal Id"
          },
          "proposal_kind": {
            "type": "string",
            "enum": [
              "eat",
              "drink",
              "adjust_generation",
              "adjust_water_production",
              "reserve_crew_for_refill",
              "request_plant_review"
            ],
            "title": "Proposal Kind"
          },
          "target_crew_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Crew Ids"
          },
          "proposed_changes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Proposed Changes"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "constraints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Constraints"
          },
          "coordination_needed": {
            "type": "boolean",
            "title": "Coordination Needed"
          },
          "feasibility": {
            "type": "string",
            "enum": [
              "verified_for_audited_scope",
              "requires_core_coordination",
              "unverified",
              "infeasible"
            ],
            "title": "Feasibility"
          },
          "rule_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rule Ids"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Ids"
          },
          "audit_result_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audit Result Id"
          },
          "expected_effect": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Effect"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "proposal_id",
          "proposal_kind",
          "target_crew_ids",
          "proposed_changes",
          "reason",
          "constraints",
          "coordination_needed",
          "feasibility",
          "rule_ids",
          "evidence_ids",
          "audit_result_id",
          "expected_effect"
        ],
        "title": "Recommendation"
      },
...
      "Resources": {
        "properties": {
          "water": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Water"
          },
          "oxygen": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Oxygen"
          },
          "food": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Food"
          },
          "power": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Power"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "water",
          "oxygen",
          "food",
          "power"
        ],
        "title": "Resources"
      },
      "RetrievalInfo": {
        "properties": {
          "requested_mode": {
            "type": "string",
            "title": "Requested Mode"
          },
          "actual_mode": {
            "type": "string",
            "enum": [
              "dense",
              "bm25",
              "mixed",
              "none",
              "not_used"
            ],
            "title": "Actual Mode"
          },
          "fallback_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Reason"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "dimensions": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dimensions"
          },
          "ranking_method": {
            "type": "string",
            "title": "Ranking Method",
            "default": "none"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requested_mode",
          "actual_mode",
          "fallback_reason",
          "model",
          "dimensions"
        ],
        "title": "RetrievalInfo"
      },
...
      "Risk": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "severity": {
            "type": "string",
            "enum": [
              "critical",
              "warning",
              "unknown",
              "info"
            ],
            "title": "Severity"
          },
          "impact_domain": {
            "type": "string",
            "enum": [
              "human",
              "public",
              "plant_dependency",
              "integration"
            ],
            "title": "Impact Domain"
          },
          "entity_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Id"
          },
          "resource": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "tick_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tick Offset"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "trigger_values": {
            "additionalProperties": true,
            "type": "object",
            "title": "Trigger Values"
          },
          "rule_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rule Ids"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "severity",
          "impact_domain",
          "entity_id",
          "resource",
          "phase",
          "tick_offset",
          "description",
          "rule_ids",
          "evidence_ids"
        ],
        "title": "Risk"
      },
...
    }
  }
}